requirements.txt
attrs==21.1.0
Automat==20.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudinary==1.26.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
emoji==1.2.0
Flask==1.1.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==1.1.1
parsel==1.6.0
PDFNetPython3==8.1.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyDispatcher==2.0.5
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-decouple==3.4
pytz==2021.1
queuelib==1.6.1
requests==2.25.1
Scrapy==2.5.0
selenium==3.141.0
service-identity==18.1.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
twilio==6.58.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
urllib3==1.26.4
w3lib==1.22.0
Werkzeug==1.0.1
zope.interface==5.4.0
